remove services/db
