+++ ee/tabby-db/src/lib.rs
@@ -1,3 +1,7 @@
+pub use invitations::InvitationDAO;
+pub use job_runs::JobRunDAO;
+pub use users::UserDAO;
+
 mod invitations;
 mod job_runs;
 mod path;
@@ -6,7 +10,6 @@ mod users;
 
 use anyhow::Result;
 use include_dir::{include_dir, Dir};
-pub use job_runs::JobRunDAO;
 use lazy_static::lazy_static;
 use rusqlite::params;
 use rusqlite_migration::AsyncMigrations;
@@ -25,7 +28,7 @@ pub struct DbConn {
 }
 
 impl DbConn {
-    #[cfg(test)]
+    #[cfg(any(test, feature = "testutils"))]
     pub async fn new_in_memory() -> Result<Self> {
         let conn = Connection::open_in_memory().await?;
         DbConn::init_db(conn).await
@@ -160,11 +163,11 @@ mod tests {
     }
 }
 
-#[cfg(test)]
-mod testutils {
+#[cfg(any(test, feature = "testutils"))]
+pub mod testutils {
     use super::*;
 
-    pub(crate) async fn create_user(conn: &DbConn) -> i32 {
+    pub async fn create_user(conn: &DbConn) -> i32 {
         let email: &str = "test@example.com";
         let password: &str = "123456789";
         conn.create_user(email.to_string(), password.to_string(), true)
