+++ ee/tabby-webserver/Cargo.toml
@@ -14,40 +14,30 @@ bincode = "1.3.3"
 chrono = { version = "0.4", features = ["serde"] }
 futures.workspace = true
 hyper = { workspace = true, features=["client"]}
-include_dir = "0.7.3"
 jsonwebtoken = "9.1.0"
 juniper.workspace = true
 juniper-axum = { path = "../../crates/juniper-axum" }
 lazy_static = "1.4.0"
 mime_guess = "2.0.4"
 pin-project = "1.1.3"
-rusqlite = { version = "0.30.0", features = ["bundled", "chrono"] }
-# `alpha-async-tokio-rusqlite` is only available from 1.1.0-alpha.2, will bump up version when it's stable
-rusqlite_migration = { version = "1.1.0-beta.1", features = ["alpha-async-tokio-rusqlite", "from-directory"] }
 rust-embed = "8.0.0"
 serde.workspace = true
 serde_json.workspace = true
 tabby-common = { path = "../../crates/tabby-common" }
+tabby-db = { path = "../../ee/tabby-db" }
 tarpc = { version = "0.33.0", features = ["serde-transport"] }
 thiserror.workspace = true
 tokio = { workspace = true, features = ["fs"] }
 tokio-cron-scheduler = "0.9.4"
-tokio-rusqlite = "0.5.0"
 tokio-tungstenite = "0.20.1"
 tower = { version = "0.4", features = ["util"] }
 tower-http = { version = "0.4.0", features = ["fs", "trace"] }
 tracing.workspace = true
 unicase = "2.7.0"
 validator = { version = "0.16.1", features = ["derive"] }
-
-[dependencies.uuid]
-version = "1.3.3"
-features = [
-    "v4",                # Lets you generate random UUIDs
-    "fast-rng",          # Use a faster (but still sufficiently random) RNG
-    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
-]
+uuid.workspace = true
 
 [dev-dependencies]
 assert_matches = "1.5.0"
 tokio = { workspace = true, features = ["macros", "process"] }
+tabby-db = { path = "../../ee/tabby-db", features = ["testutils"] }
