+++ ee/tabby-webserver/src/lib.rs
@@ -1,6 +1,5 @@
 mod handler;
 mod hub;
-mod path;
 mod repositories;
 mod schema;
 mod service;
